Raspberry Pi 4 HDMI workaround
Just using vcgencmd to turn HDMI on and off does not seem to work - if
output is turned completely off (as with tvservice -o, which seems to
happen automatically after some time), vcgencmd can turn the display on
but yields no picture.

This commit introduces a workaround: use tvservice to turn on HDMI,
then fbset to actually get the output on. For some reason we end up in
tty1, though. Finally use chvt to switch to tty7.